requirements.txt
Twisted>=12.0.0
PyYAML
